Added abstraction for making it simple to control how WebHooks are sent. Now it is easy to enqueue messages into something like an Azure Storage Queue and also to control the format of what the actual WebHooks messages look like. This means that if you want the message format to be different then it's now easy to do.
Added support for sending WebHooks directly to an Azure Storage Queue and also a mechanism for dequeing messages from the queue and then send out the WebHooks as normal. That allows for implementing highly scalable systems where the outgoing WebHooks are persisted in a queue and then can be sent from any number of queue client that poll for messages as they become available.